MueLu: implementation of uncoupled structured aggregation.
This adds an important feature to the structured aggregation capability.
Aggregates do not cross processor boundaries and do not need knowledge
 of the global layout of the mesh, only local data is used.
The main additions are the UncoupledIndexManager, the coupled flag in
 MueLu_StructuredAggregationFactory and the structured aggregation flag
 in the MueLu_CoordinatesTransferFactory.
The new aggregation scheme has been tested in 1, 2 and 3D and the 3D
algorithm is tested on a 4 level hierarchy.