MueLu: structured region driver
Fixing some bugs by reverting back to scalar type for the calculation of coarse grid composite map. This is fragile and will eventually need to be removed but lets keep things as they are for now.
A minor change was done to go from SC to Teuchos::ScalarTraits<SC>::magnitudeType which fixes the build when std::complex<> is enabled.