MueLu: structured region driver, fix issue in regionalToComposite
There is something wrong in Epetra_IntVector when it does an export.
I cannot catch it so I reverted to using Epetra_Vector instead...
The structured region driver makes progress torward running a full simulation.
Currently difficulties reside in makeCoarseLevelMaps but it seems to be logic issues not implementation one so rather fixable...